sentiment analysis initializing
