+++ applications/sentiment_analysis/README.md
@@ -12,9 +12,10 @@
 
 > 这个薯片味道真的太好了，口感很脆，只是包装很一般。
 
-可以看到，顾客在口感、包装和味道 三个维度上对薯片进行了评价，顾客在味道和口感两个方面给出了好评，但是在包装上给出了负面的评价。只有通过这种比较细粒度的分析，商家才能更有针对性的发现问题，进而改进自己的产品或服务。
+可以看到，顾客在口感、包装和味道 三个维度上对薯片进行了评价，顾客在味道和口感两个方面给出了好评，但是在包装上给出了负面的评价。由此看来，针对一个产品或服务，用户可从多个维度进行评论，每个维度都会有相应的观点。
 
-本项目基于这样的考量，提供出一套完整的细粒度情感分析解决方案，期望能够在评论语句中评论维度的粒度级别进行情感分析。
+基于这样的考虑，本项目提出了一套评论观点抽取的情感分析能力，即对于给定文本，首先会抽取该文本中的评论维度和观点，然后分析不同维度的情感极性。
+
 
 ## 2. 产品功能介绍
 
@@ -22,7 +23,7 @@
 为了降低技术门槛，方便开发者共享效果领先的情感分析技术，PaddleNLP本次开源的情感分析系统，具备三大亮点：
 
 - 覆盖任务全
-    - 集成句子级情感分类、评论观点抽取、属性级情感分类等多种情感分析能力，并开源模型，且打通模型训练、评估、预测部署全流程。
+    - 集成评论观点抽取、属性级情感分类等情感分析能力，并开源模型，且打通模型训练、评估、预测部署全流程。
 - 效果领先
     - 集成百度研发的基于情感知识增强的预训练模型SKEP，为各类情感分析任务提供统一且强大的情感语义表示能力。
 - 预测性能强
@@ -30,7 +31,7 @@
 
 ### 2.2 架构&功能
 
-针对以上提到的细粒度情感分析，我们提出的解决方案如下图所示。整个情感分析的过程包含两个阶段，依次是评论维度和观点抽取模型，细粒度情感分类模型。对于给定的一段文本，首先基于前者抽取出文本语句中潜在的评论维度以及该维度相应的评论观点，然后将评论维度、观点以及原始文本进行拼接，传给细粒度情感分类模型以识别出该评论维度的情感色彩。
+以上提到的解决方案如图1所示，整个情感分析的过程大致包含两个阶段，依次是评论维度和观点抽取模型，细粒度情感分类模型。对于给定的一段文本，首先基于前者抽取出文本语句中潜在的评论维度以及该维度相应的评论观点，然后将评论维度、观点以及原始文本进行拼接，传给细粒度情感分类模型以识别出该评论维度的情感色彩。
 
 这里需要提到的是，由于目前市面上的大多数模型是基于通用语料训练出来的，这些模型可能并不会对情感信息那么敏感。基于这样的考量，本项目使用了百度自研的 SKEP 预训练模型，其在预训练阶段便设计了多种情感信息相关的预训练目标进行训练。作为一种情感专属的模型，其更适合用来做上边提到的评论维度和观点抽取任务，以及细粒度情感分类任务。
