+++ applications/sentiment_analysis/classification/run_train.sh
@@ -1,4 +1,4 @@
-export CUDA_VISIBLE_DEVICES=1
+export CUDA_VISIBLE_DEVICES=6
 
 python  train.py \
         --train_path "../data/cls_data/train.txt" \
