+++ applications/sentiment_analysis/demo.py
