+++ applications/sentiment_analysis/pp_minilm/run_evaluate.sh
@@ -1,8 +1,8 @@
 export CUDA_VISIBLE_DEVICES=1
 
 python  evaluate.py \
-        --base_model_path "../checkpoints/ppminilm" \
-        --model_path "../checkpoints/sp_checkpoints/best.pdparams" \
+        --base_model_path "../checkpoints/pp_minilm" \
+        --model_path "../checkpoints/pp_checkpoints/best.pdparams" \
         --test_path "../data/cls_data/test.txt" \
         --label_path "../data/cls_data/label.dict" \
         --batch_size 16 \
