+++ applications/sentiment_analysis/pp_minilm/run_performance_test.sh
@@ -1,8 +1,8 @@
 export CUDA_VISIBLE_DEVICES=1
 
-python  static_predict.py \
-        --base_model_path "../checkpoints/ppminilm" \
-        --model_path "../checkpoints/sp_checkpoints/quant/infer" \
+python  performance_test.py \
+        --base_model_path "../checkpoints/pp_minilm" \
+        --model_path "../checkpoints/pp_checkpoints/quant/infer" \
         --test_path "../data/cls_data/test.txt" \
         --label_path "../data/cls_data/label.dict" \
         --num_epochs 10 \
