+++ applications/sentiment_analysis/utils.py
@@ -15,8 +15,84 @@
 import paddle
 import random
 import numpy as np
+from seqeval.metrics.sequence_labeling import get_entities
+
 
 def set_seed(seed):
     paddle.seed(seed)
     random.seed(seed)
     np.random.seed(seed)
+
+def load_dict(dict_path):
+    with open(dict_path, "r", encoding="utf-8") as f:
+        words = [word.strip() for word in f.readlines()]
+        word2id = dict(zip(words, range(len(words))))
+        id2word = dict((v, k) for k, v in word2id.items())
+
+        return word2id, id2word
+
+def read_test_file(data_path):
+    with open(data_path, "r", encoding="utf-8") as f:
+        for line in f.readlines():
+            line = line.strip().replace(" ", "")
+            yield {"text": line}
+
+def decoding(text, tag_seq):
+    assert len(text) == len(
+        tag_seq), f"text len: {len(text)}, tag_seq len: {len(tag_seq)}"
+
+    puncs = list(",.?;!，。？；！")
+    splits = [idx for idx in range(len(text)) if text[idx] in puncs]
+
+    prev = 0
+    sub_texts, sub_tag_seqs = [], []
+    for i, split in enumerate(splits):
+        sub_tag_seqs.append(tag_seq[prev:split])
+        sub_texts.append(text[prev:split])
+        prev = split
+    sub_tag_seqs.append(tag_seq[prev:])
+    sub_texts.append((text[prev:]))
+
+    ents_list = []
+    for sub_text, sub_tag_seq in zip(sub_texts, sub_tag_seqs):
+        ents = get_entities(sub_tag_seq, suffix=False)
+        ents_list.append((sub_text, ents))
+
+    aps = []
+    no_a_words = []
+    for sub_tag_seq, ent_list in ents_list:
+        sub_aps = []
+        sub_no_a_words = []
+        for ent in ent_list:
+            ent_name, start, end = ent
+            if ent_name == "Aspect":
+                aspect = sub_tag_seq[start:end + 1]
+                sub_aps.append([aspect])
+                if len(sub_no_a_words) > 0:
+                    sub_aps[-1].extend(sub_no_a_words)
+                    sub_no_a_words.clear()
+            else:
+                ent_name == "Opinion"
+                opinion = sub_tag_seq[start:end + 1]
+                if len(sub_aps) > 0:
+                    sub_aps[-1].append(opinion)
+                else:
+                    sub_no_a_words.append(opinion)
+
+        if sub_aps:
+            aps.extend(sub_aps)
+            if len(no_a_words) > 0:
+                aps[-1].extend(no_a_words)
+                no_a_words.clear()
+        elif sub_no_a_words:
+            if len(aps) > 0:
+                aps[-1].extend(sub_no_a_words)
+            else:
+                no_a_words.extend(sub_no_a_words)
+
+    if no_a_words:
+        no_a_words.insert(0, "None")
+        aps.append(no_a_words)
+
+    return aps
+
